backend/requirements.txt

numpy>=1.20.0
tensorflow>=2.7.0
opencv-python>=4.5.5
matplotlib>=3.5.1
pillow>=9.0.0
scikit-learn>=1.0.2
flask>=2.0.2
flask-cors>=3.0.10
python-dotenv>=0.19.2
